# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个基于Go语言开发的任务调度系统API文档",
        "title": "GoDo任务调度系统API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "description": "用户登录接口，验证用户名和密码，返回JWT令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户认证"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "登录请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登录成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.LoginResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "登录失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/add_shell_task": {
            "post": {
                "description": "创建一个新的Shell任务，支持定时执行",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "添加Shell任务",
                "parameters": [
                    {
                        "description": "任务创建参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.AddShellTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.AddShellTaskResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/delete": {
            "delete": {
                "description": "根据任务ID和用户名删除指定任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "删除任务",
                "parameters": [
                    {
                        "description": "删除任务参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.DeleteTaskRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "删除任务失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/list/{name}": {
            "get": {
                "description": "根据用户名获取该用户的所有任务",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "获取用户任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"admin\"",
                        "description": "用户名",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.ListTaskResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/tasks/upload_script": {
            "post": {
                "description": "上传脚本文件到服务器，用于后续任务执行",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "任务管理"
                ],
                "summary": "上传脚本文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "脚本文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "上传成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/controller.UploadScriptResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "文件未上传",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "文件保存失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.AddShellTaskRequest": {
            "description": "添加Shell任务的请求参数",
            "type": "object",
            "required": [
                "command",
                "description",
                "owner_name",
                "scheduled_time",
                "task_name",
                "timeout",
                "use_shell"
            ],
            "properties": {
                "args": {
                    "description": "命令参数",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "backup.sh",
                        "--full"
                    ]
                },
                "command": {
                    "description": "执行命令",
                    "type": "string",
                    "example": "/bin/bash"
                },
                "description": {
                    "description": "任务描述",
                    "type": "string",
                    "example": "每日数据备份任务"
                },
                "owner_name": {
                    "description": "任务所有者",
                    "type": "string",
                    "example": "admin"
                },
                "scheduled_time": {
                    "description": "Cron表达式",
                    "type": "string",
                    "example": "0 2 * * *"
                },
                "task_name": {
                    "description": "任务名称",
                    "type": "string",
                    "example": "daily-backup"
                },
                "timeout": {
                    "description": "超时时间(秒)，最大2小时",
                    "type": "integer",
                    "maximum": 7200,
                    "example": 1800
                },
                "use_shell": {
                    "description": "是否使用Shell",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "controller.AddShellTaskResponseData": {
            "description": "添加任务成功响应数据",
            "type": "object",
            "properties": {
                "task_id": {
                    "description": "新创建的任务ID",
                    "type": "integer",
                    "example": 12345
                }
            }
        },
        "controller.DeleteTaskRequest": {
            "description": "删除任务的请求参数",
            "type": "object",
            "properties": {
                "task_id": {
                    "description": "任务ID",
                    "type": "integer",
                    "example": 12345
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "controller.ListTaskResponseData": {
            "description": "任务列表响应数据结构",
            "type": "object",
            "properties": {
                "tasks": {
                    "description": "任务列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/scheduler.Task"
                    }
                }
            }
        },
        "controller.LoginRequest": {
            "description": "用户登录请求参数",
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "password123"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
        "controller.LoginResponseData": {
            "description": "登录成功响应数据",
            "type": "object",
            "properties": {
                "token": {
                    "description": "JWT令牌",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
        "controller.UploadScriptResponseData": {
            "description": "脚本上传成功响应数据",
            "type": "object",
            "properties": {
                "file_name": {
                    "description": "上传后的文件名",
                    "type": "string",
                    "example": "1699123456789-script.sh"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "scheduler.Task": {
            "type": "object"
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}